requirements.txt
python-telegram-bot==20.7
Pillow==10.3.0
qrcode==7.4.2